+++ src/Game/GameObjects/Water.cpp
@@ -28,7 +28,7 @@ void Water::render() const
 	renderBlock(EBlockLocation::TopRight);
 }
 
-void Water::update(const uint64_t delta)
+void Water::update(const double delta)
 {
 	m_spriteAnimator.update(delta);
 }
