+++ src/Game/GameObjects/Water.h
@@ -26,7 +26,7 @@ public:
 
 	Water(const glm::vec2& position, const glm::vec2& size, const float rotation, const float layer);
 	virtual void render() const override;
-	void update(const uint64_t) override;
+	void update(const double) override;
 
 private:
 	void renderBlock(const EBlockLocation eBricklocation) const;
