+++ src/Renderer/Sprite.cpp
@@ -104,7 +104,7 @@ namespace RenderEngine
 		m_framesDescription = std::move(framesDescriptions);
 	}
 
-	uint64_t Sprite::getFramesDuration(const size_t frameId) const
+	double Sprite::getFramesDuration(const size_t frameId) const
 	{
 		return m_framesDescription[frameId].duration;
 	}
